+++ src/index.ts
@@ -1,7 +1,5 @@
 import {
   AsyncPredicate,
-  Mapper,
-  AsyncMapper,
   Page,
   Pager,
   FluentIterable,
@@ -9,7 +7,7 @@ import {
   FluentGroup,
 } from './types';
 import './types-key';
-import { OrderAssurable } from './assure-order-types';
+import { OrderAssurable } from './types/assure-order-types';
 import {
   Predicate,
   Reducer,
@@ -19,7 +17,7 @@ import {
   AsyncAction,
   Group,
   FluentEmitter,
-} from './types-base';
+} from './types/base';
 
 import fluent from './fluent';
 import { fluentAsync, fluentEmit } from './fluent-async';
@@ -27,6 +25,7 @@ import depaginate from './depaginator';
 import { interval } from './interval';
 import { mergeIterators } from './async-base/merging';
 import { assureOrder as o, assureOrderDescending as od } from './utils';
+import { AsyncMapper, Mapper } from 'augmentative-iterable';
 
 export * from './extend-async';
 export * from './extend';
