+++ src/recipes/emit-recipe.ts
@@ -1,6 +1,6 @@
 import { AnyIterable } from 'augmentative-iterable';
 import { EventEmitter } from 'events';
-import { FluentEmitter } from '../types-base';
+import { FluentEmitter } from '../types/base';
 import { ResolverIngredient } from './ingredients';
 
 export function emitRecipe(forEach: ResolverIngredient) {
