+++ src/recipes/reduce-recipe.ts
@@ -1,6 +1,6 @@
 import { identity } from '../utils';
 import { AnyIterable } from 'augmentative-iterable';
-import { AsyncReducer } from '../types-base';
+import { AsyncReducer } from '../types/base';
 import { ReduceAndMapIngredient } from './ingredients';
 
 export function reduceRecipe(reduceAndMap: ReduceAndMapIngredient) {
