+++ src/recipes/wait-all-recipe.ts
@@ -1,4 +1,3 @@
-import { AsyncMapper } from '../types';
 import { AnyIterable } from 'augmentative-iterable';
 import { BasicIngredients } from './ingredients';
 import { AnyMapper } from '../types-internal';
